+++ src/commonTest/kotlin/org/hisp/dhis/lib/expression/function/VectorAvgTest.kt
@@ -8,7 +8,7 @@ import kotlin.test.assertEquals
  *
  * @author Jan Bernitt
  */
-internal class AvgTest : AbstractAggregateFunctionTest() {
+internal class VectorAvgTest : AbstractVectorBasedTest() {
 
     @Test
     fun testAvg_Single() {
