+++ src/commonTest/kotlin/org/hisp/dhis/lib/expression/function/VectorMinTest.kt
@@ -8,7 +8,7 @@ import kotlin.test.assertEquals
  *
  * @author Jan Bernitt
  */
-internal class MinTest : AbstractAggregateFunctionTest() {
+internal class VectorMinTest : AbstractVectorBasedTest() {
 
     @Test
     fun testMin() {
